requirements.txt
scapy~=2.6.1
pyyaml~=6.0.3